% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{commune}
\alias{commune}
\title{Administrative level boundary 3 (commune borders) for Comoros Union}
\format{
A SpatialPolygonsDataFrame with 55 features:
\describe{
  \item{\code{Shape_Leng}}{Shape length}
  \item{\code{Shape_Area}}{Shape area}
  \item{\code{ADM3_EN}}{Name of prefecture (English)}
  \item{\code{ADM3_PCODE}}{Postal code of commune}
  \item{\code{ADM3_REF}}{Commune reference}
  \item{\code{ADM3ALT1EN}}{First alternative commune name (English)}
  \item{\code{ADM3ALT2EN}}{Second alternative commune name (English)}
  \item{\code{ADM2_EN}}{Name of prefecture (English)}
  \item{\code{ADM2_PCODE}}{Postal code of prefecture}
  \item{\code{ADM1_EN}}{Name of island (English)}
  \item{\code{ADM1_PCODE}}{Postal code of island}
  \item{\code{ADM0_EN}}{Name of country (English)}
  \item{\code{ADM0_PCODE}}{Postal code of country}
  \item{\code{date}}{Date produced}
  \item{\code{validOn}}{Date map is valid from}
  \item{\code{validTo}}{Date map is valid to}
}
}
\source{
\url{https://data.humdata.org/dataset/comoros-administrative-level-0-3-boundaries}
}
\usage{
commune
}
\description{
Administrative level boundary 3 (commune borders) for Comoros Union
}
\examples{
commune

}
\keyword{datasets}
