% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{country}
\alias{country}
\title{Administrative level boundary 0 (country borders) for Comoros Union}
\format{A SpatialPolygonsDataFrame with 1 feature:
\describe{
  \item{\code{Shape_Leng}}{Shape length}
  \item{\code{Shape_Area}}{Shape area}
  \item{\code{ADM0_EN}}{Name of country (English)}
  \item{\code{ADM0_PCODE}}{Postal code of country}
  \item{\code{ADM0_REF}}{Country reference}
  \item{\code{ADM0ALT1EN}}{First alternative country name (English)}
  \item{\code{ADM0ALT2EN}}{Second alternative country name (English)}
  \item{\code{date}}{Date produced}
  \item{\code{validOn}}{Date map is valid from}
  \item{\code{validTo}}{Date map is valid to}
}}
\source{
\url{https://data.humdata.org/dataset/comoros-administrative-level-0-3-boundaries}
}
\usage{
country
}
\description{
Administrative level boundary 0 (country borders) for Comoros Union
}
\examples{
country

}
\keyword{datasets}
