% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{island}
\alias{island}
\title{Administrative level boundary 1 (island borders) for Comoros Union}
\format{
A SpatialPolygonsDataFrame with 3 features:
\describe{
  \item{\code{Shape_Leng}}{Shape length}
  \item{\code{Shape_Area}}{Shape area}
  \item{\code{ADM1_EN}}{Name of island (English)}
  \item{\code{ADM1_PCODE}}{Postal code of island}
  \item{\code{ADM1_REF}}{Island reference}
  \item{\code{ADM1ALT1EN}}{First alternative island name (English)}
  \item{\code{ADM1ALT2EN}}{Second alternative island name (English)}
  \item{\code{ADM0_EN}}{Name of country (English)}
  \item{\code{ADM0_PCODE}}{Postal code of country}
  \item{\code{date}}{Date produced}
  \item{\code{validOn}}{Date map is valid from}
  \item{\code{validTo}}{Date map is valid to}
}
}
\source{
\url{https://data.humdata.org/dataset/comoros-administrative-level-0-3-boundaries}
}
\usage{
island
}
\description{
Administrative level boundary 1 (island borders) for Comoros Union
}
\examples{
island

}
\keyword{datasets}
