% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ppl}
\alias{ppl}
\title{Populated places (points) map for Comoros Union}
\format{
A SpatialPointsDataFrame with 144 features:
\describe{
  \item{\code{osm_id}}{Open Street Map identifier}
  \item{\code{is_in}}{Where is this place located?}
  \item{\code{source}}{Source of populated place information}
  \item{\code{place}}{Type of populated place}
  \item{\code{population}}{Population number}
  \item{\code{name}}{Name of populated place}
}
}
\source{
\url{https://data.humdata.org/dataset/hotosm_com_populated_places}
}
\usage{
ppl
}
\description{
Populated places (points) map for Comoros Union
}
\examples{
ppl

}
\keyword{datasets}
