% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prefecture}
\alias{prefecture}
\title{Administrative level boundary 2 (prefecture borders) for Comoros Union}
\format{A SpatialPolygonsDataFrame with 17 features:
\describe{
  \item{\code{Shape_Leng}}{Shape length}
  \item{\code{Shape_Area}}{Shape area}
  \item{\code{ADM2_EN}}{Name of prefecture (English)}
  \item{\code{ADM2_PCODE}}{Postal code of prefecture}
  \item{\code{ADM2_REF}}{Prefecture reference}
  \item{\code{ADM2ALT1EN}}{First alternative prefecture name (English)}
  \item{\code{ADM2ALT2EN}}{Second alternative prefecture name (English)}
  \item{\code{ADM1_EN}}{Name of island (English)}
  \item{\code{ADM1_PCODE}}{Postal code of island}
  \item{\code{ADM0_EN}}{Name of country (English)}
  \item{\code{ADM0_PCODE}}{Postal code of country}
  \item{\code{date}}{Date produced}
  \item{\code{validOn}}{Date map is valid from}
  \item{\code{validTo}}{Date map is valid to}
}}
\source{
\url{https://data.humdata.org/dataset/comoros-administrative-level-0-3-boundaries}
}
\usage{
prefecture
}
\description{
Administrative level boundary 2 (prefecture borders) for Comoros Union
}
\examples{
prefecture

}
\keyword{datasets}
